Add files via upload
- 新增Prometheus用于观察服务端三个粒度的相关统计信息(QPS/error_rate/rt)
- 修复了流量治理部分的部分bug
- 经测试能够正常地推送流量治理规则 能够正常地通过本地Prometheus观测请求处理指标

+++ src/main/java/com/example/rpc/stability/system/SystemRuleManager.java
@@ -68,7 +68,7 @@ public class SystemRuleManager {
         rule.setHighestCpuUsage(meta.highestCpuUsage);
         rule.setQps(meta.maxQps);
         rule.setMaxThread(meta.maxThread);
-        rule.setAvgRt(meta.maxAvgRt);
+        rule.setAvgRt((long) meta.maxAvgRt);
         rules.add(rule);
         // 推送到Sentinel，复用你的ZK推送器
         SentinelRuleZkManager.pushSystemRules(rules, System.currentTimeMillis());
